Revert "Feature/observables"
