+++ index.ts
@@ -1,3 +1,1 @@
-export { BugSplatErrorHandler } from "./src/bugsplat-error-handler";
-export { BugSplat } from "./src/bugsplat";
-export { BugSplatConfig, BugSplatConfigToken } from "./src/bugsplat-config";
+export { BugSplatErrorHandler, BugSplat, BugSplatConfig, BugSplatConfigToken } from "./src/bugsplat-error-handler";
