+++ src/bugsplat-error-handler.spec.ts
@@ -1,16 +1,10 @@
 import { TestBed, async } from '@angular/core/testing';
 import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { HttpClientModule, HttpClient } from "@angular/common/http";
-import { XHRBackend, ResponseOptions, Http, BaseRequestOptions } from "@angular/http";
+import { XHRBackend, ResponseOptions } from "@angular/http";
+import { BugSplat } from "./bugsplat-error-handler";
 import { MockBackend } from '@angular/http/testing';
 import { HttpErrorResponse } from '@angular/common/http';
-import { BugSplat } from '../src/bugsplat';
-import { BugSplatPostEventType } from '../src/bugsplat-post-event';
-import { TestBedInitializer } from './init';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/observable/throw';
 
 const testUser = "Fred";
 const testPassword = "Flintstone";
@@ -18,17 +12,21 @@ const testDatabase = "octomore"
 
 describe('BugSplat', () => {
 
-    let TestBed;
-
     beforeAll(() => {
-        TestBed = TestBedInitializer.getTestBed();
+        TestBed.initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting());
     });
-    
+
     beforeEach(() => TestBed.configureTestingModule({
-        imports: [HttpClientTestingModule]
+        imports: [HttpClientModule],
+        providers: [
+            HttpClient,
+            { provide: XHRBackend, useClass: MockBackend }
+        ]
     }));
 
-    it('should publish an event on post success', async(() => {
+    it('should pass response data to callback', async(() => {
+        const http = TestBed.get(HttpClient);
+        const mockBackend = TestBed.get(XHRBackend);
         const mockSuccessResponse = {
             status: 'success',
             current_server_time: 1505832461,
@@ -39,47 +37,31 @@ describe('BugSplat', () => {
             appName: "bugsplat-ng4-tests",
             appVersion: "1.0.0.0",
             database: testDatabase
-        };
-        const http = TestBed.get(HttpClient);
-        http.post = (url, body) => {
-            return Observable.of(mockSuccessResponse);
-        };
+        }
         const bugsplat = new BugSplat(config, http);
-        bugsplat.getObservable().subscribe(event => {
-            expect(event.type).toEqual(BugSplatPostEventType.Success);
-            expect(event.responseData.message).toEqual("Crash successfully posted");
-            expect(event.responseData.success).toEqual(true);
-            expect(event.responseData.crash_id).toMatch(/\d{1,}/);
-        }, err => {
-            throw err;
+        setMockBackendSuccessResponse(mockBackend, mockSuccessResponse);
+        bugsplat.setCallback((err, data, context) => {
+            expect(data.message).toEqual("Crash successfully posted");
+            expect(data.status).toEqual("success");
+            expect(data.crash_id).toMatch(/\d{1,}/);
         });
         bugsplat.post(new Error("foobar!"));
-    }));
+    }), 15000);
 
-    it('should publish an event on post error', async(() => {
+    it('should pass response error to callback', async(() => {
+        const http = TestBed.get(HttpClient);
+        const mockBackend = TestBed.get(XHRBackend);
         const mockFailureStatus = 400;
-        const mockFailureResponse = new HttpErrorResponse({
-            status: mockFailureStatus,
-            statusText: 'Bad Request',
-            url: 'https://octomore.bugsplat.com/post/js/',
-            error: null
-        })
         const config = {
             appName: "",
             appVersion: "",
             database: testDatabase
-        };
-        const http = TestBed.get(HttpClient);
-        http.post = (url, body) => {
-            return Observable.throw(mockFailureResponse);
-        };
+        }
         const bugsplat = new BugSplat(config, http);
-        bugsplat.getObservable().subscribe(event => {
-            expect(event.type).toEqual(BugSplatPostEventType.Error);
-            expect(event.responseData.success).toBe(false);
-            expect(event.responseData.message).toContain("400 Bad Request");
-        }, err => {
-            throw err;
+        setMockBackendFailureResponse(mockBackend, mockFailureStatus);
+        bugsplat.setCallback((err, data, context) => {
+            expect(err).not.toBe(null);
+            expect(err.status).toEqual(mockFailureStatus);
         });
         bugsplat.post(new Error("foobar!"));
     }));
@@ -100,5 +82,22 @@ describe('BugSplat', () => {
         const expectedMessage = "BugSplat Error: Could not add file " + file.name + ". Upload bundle size limit exceeded!";
         bugsplat.addAddtionalFile(file);
         expect(spy).toHaveBeenCalledWith(expectedMessage);
-    }));
+        }));
+
+    function setMockBackendSuccessResponse(mockBackend, body) {
+        mockBackend.connections.subscribe((connection) => {
+            connection.mockRespond(new Response(
+                new ResponseOptions({
+                    body: JSON.stringify(body)
+                })));
+        });
+    }
+
+    function setMockBackendFailureResponse(mockBackend, status) {
+        mockBackend.connections.subscribe((connection) => {
+            connection.mockFailureResponse(new HttpErrorResponse({
+                status: status
+            }));
+        });
+    }
 });
